[Fix] #112 - TextField 디자인시스템 레거시 파일 삭제 및 파일명 변경
