+++ Winey/Common/DesignSystem/Sources/DesignSystem/Sources/TextField/WITextFieldView.swift
@@ -1,310 +1,321 @@
-////
-////  File.swift
-////
-////
-////  Created by 김응관 on 2023/07/17.
-////
 //
-//import Combine
-//import UIKit
+//  File.swift
+//  
 //
-//import SnapKit
+//  Created by 김응관 on 2023/08/17.
 //
-//public final class WITextFieldView: UIView {
-//
-//    // MARK: - Properties
-//
-//    /// uploadPrice: 입력된 금액을 UploadViewController로 전달해주는 클로져
-//    /// unit: "원", "일" 등 텍스트필드 뒤에 붙을 단위를 표시해주는 label 객체
-//    /// textField: 금액및 일수를 입력하는 텍스트필드
-//    /// price: textField의 text에 들어갈 문자열
-//    /// label: textField의 숫자 오른쪽에 들어갈 단위의 타입을 담은 변수
-//    public var uploadPrice: ((_ data: Int) -> Void)?
-//
-//    private let unitLabel: UILabel = UILabel()
-//    private let textField: UITextField = UITextField()
-//
-//    public let pricePublisher = PassthroughSubject<Int, Never>()
-//    public let textFieldDidEndEditingPublisher = PassthroughSubject<Void, Never>()
-//
-//    private var price: Int = 0 {
-//        didSet { pricePublisher.send(price) }
-//    }
-//
-//    public var label: Unit?
-//
-//    public let textLength: Int?
-//
-//    public func makeErrorView() {
-//        textField.makeBorder(width: Const.borderWidth, color: Const.errorColor)
-//        textField.textColor = Const.errorColor
-//    }
-//
-//    public func makeActiveView() {
-//        textField.makeBorder(width: Const.borderWidth, color: Const.activeColor)
-//        textField.textColor = Const.activeColor
-//    }
-//
-//    public func makeInactiveView() {
-//        textField.makeBorder(width: Const.borderWidth, color: Const.inactivateBorderColor)
-//        textField.textColor = Const.inactivateTextColor
-//    }
-//
-//    public override var intrinsicContentSize: CGSize {
-//        CGSize(width: UIScreen.main.bounds.width, height: Const.textFieldHeight)
-//    }
-//
-//    public init(price: String? = nil, label: Unit? = nil, textLength: NumberType? = nil) {
-//        self.label = label
-//        self.textLength = textLength?.number
-//        super.init(frame: .zero)
-//        setUI()
-//        setLayout()
-//    }
-//
-//    required init?(coder: NSCoder) {
-//        fatalError("init(coder:) has not been implemented")
-//    }
-//
-//    /// become, resign: 텍스트필드가 firstResponder가 될지 말지 결정해주는 함수
-//    public func become() {
-//        textField.becomeFirstResponder()
-//    }
-//
-//    public func resign() {
-//        textField.resignFirstResponder()
-//    }
-//}
-//
-//extension WITextFieldView {
-//
-//    /// setUI: TextField의 속성값 지정
-//    private func setUI() {
-//        setAddTarget()
-//
-//        textField.delegate = self
-//        textField.placeholder = "0"
-//        textField.tintColor = Const.tintColor
-//        textField.textAlignment = .right
-//        textField.keyboardType = .numberPad
-//        textField.textColor = Const.inactivateTextColor
-//        textField.font = Typography.font(style: .headLine2, weight: .Bold)
-//        textField.addLeftPadding(width: Const.leftPadding)
-//        textField.addRightPadding(width: Const.rightPadding)
-//        textField.makeCornerRound(radius: Const.cornerRadius)
-//        textField.makeBorder(width: Const.borderWidth, color: Const.inactivateBorderColor)
-//        textField.backgroundColor = .winey_gray0
-//
-//        unitLabel.setText(label?.text, attributes: Const.labelAttributes)
-//    }
-//
-//    /// setAddTarget: 텍스트필드가 편집 중일때 액션함수 작동하게끔 설정
-//    private func setAddTarget() {
-//        textField.addTarget(self, action: #selector(textfieldDidChange), for: .editingChanged)
-//    }
-//
-//    private func setLayout() {
-//        addSubview(textField)
-//
-//        textField.snp.makeConstraints {
-//            $0.edges.equalToSuperview()
-//        }
-//
-//        addSubview(unitLabel)
-//
-//        unitLabel.snp.makeConstraints {
-//            $0.trailing.equalToSuperview().inset(18)
-//            $0.centerY.equalTo(textField.snp.centerY)
-//            $0.width.equalTo(16)
-//        }
-//    }
-//
-//    /// makeComma: 텍스트필드의 text에 콤마를 붙여주는 함수
-//    private func makeComma() {
-//        guard let str = self.textField.text else {
-//            textField.placeholder = ""
-//            return
-//        }
-//
-//        var answer: Int = 0
-//
-//        if str.count > 3 {
-//            answer = Int(str.replacingOccurrences(of: ",", with: "")) ?? 0
-//        } else {
-//            answer = Int(str) ?? 0
-//        }
-//
-//        uploadPrice?(answer)
-//
-//        self.price = answer
-//        self.textField.text = answer.addCommaToString()
-//    }
-//
-//}
-//
-//extension WITextFieldView {
-//    /// Const: 커스텀 텍스트필드 제작에 필요한 여러 레이아웃, 속성 값들을 단일 상수로 선언
-//    enum Const {
-//
-//        static let unitTrailing: CGFloat = 18.0
-//        static let textFieldHeight: CGFloat = 56.0
-//        static let labelWidth: CGFloat = 16.0
-//        static let leftPadding: CGFloat = 48.0
-//        static let rightPadding: CGFloat = 38
-//        static let cornerRadius: CGFloat = 5
-//        static let borderWidth: CGFloat = 1
-//
-//        static let tintColor = UIColor.winey_purple400
-//        static let activeColor = UIColor.winey_purple400
-//        static let errorColor = UIColor.winey_red500
-//
-//        static let unitColor = UIColor.winey_gray900
-//
-//        static let inactivateBorderColor = UIColor.winey_gray200
-//        static let inactivateTextColor = UIColor.winey_gray500
-//
-//        static let inactiveTextFieldAttributes = Typography.Attributes(
-//            style: .headLine2,
-//            weight: .bold,
-//            textColor: inactivateTextColor
-//        )
-//
-//        static let activeTextFieldAttributes = Typography.Attributes(
-//            style: .headLine2,
-//            weight: .bold,
-//            textColor: activeColor
-//        )
-//
-//        static let labelAttributes = Typography.Attributes(
-//            style: .headLine4,
-//            weight: .bold,
-//            textColor: unitColor
-//        )
-//    }
-//}
-//
-///// 텍스트필드 양옆에 여백을 주는 함수
-//extension UITextField {
-//    func addLeftPadding(width: CGFloat) {
-//        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
-//        self.leftView = paddingView
-//        self.leftViewMode = .always
-//    }
-//
-//    func addRightPadding(width: CGFloat) {
-//        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
-//        self.rightView = paddingView
-//        self.rightViewMode = .always
-//    }
-//}
-//
-//extension WITextFieldView: UITextFieldDelegate {
-//
-//    /// textFieldDidBeginEditing: textField의 편집이 시작될 때의 동작을 정의한 함수
-//    public func textFieldDidBeginEditing(_ textField: UITextField) {
-//        textField.textColor = Const.activeColor
-//        textField.makeBorder(width: Const.borderWidth, color: Const.activeColor)
-//
-//        pricePublisher.send(price)
-//        if let pure = textField.text {
-//            if pure == "0" { textField.text = nil }
-//            else { textField.text = pure }
-//        }
-//    }
-//
-//    /// textfieldDidChange: textField의 text가 변경되었을때 작동하는 함수
-//    @objc
-//    private func textfieldDidChange(_ sender: UITextField) {
-//
-//        makeComma()
-//
-//        guard let text = textField.text else { return }
-//        if text == "0" {
-//            textField.text = ""
-//        } else {
-//            textField.text = text
-//        }
-//    }
-//
-//    /// textFieldDidEndEditing: textField의 편집이 종료되었을때 작동하는 함수
-//    public func textFieldDidEndEditing(_ textField: UITextField) {
-//        if ((textField.text?.isEmpty) != nil) {
-//            textField.placeholder = "0"
-//            self.makeInactiveView()
-//        }
-//        textFieldDidEndEditingPublisher.send(Void())
-//    }
-//
-//    /// textField: 텍스트필드에 새로운 문자가 추가되었을때 text를 바꿔주는 함수
-//    public func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange,
-//                          replacementString string: String) -> Bool {
-//
-//        var price: Int = 0
-//
-//        if let pureText = textField.text {
-//            price = pureText.count + string.count - range.length
-//        }
-//        return !(price > textLength ?? 0)
-//    }
-//}
-//
-////extension UIView {
-////    func addSubviews(_ views: UIView...) {
-////        views.forEach { self.addSubview($0) }
-////    }
-////
-////    func makeCornerRound(radius: CGFloat) {
-////        layer.cornerRadius = radius
-////        layer.masksToBounds = true
-////    }
-////
-////    func makeBorder(width: CGFloat, color: UIColor ) {
-////        layer.borderWidth = width
-////        layer.borderColor = color.cgColor
-////    }
-////}
-//
-////extension Int {
-////    func addCommaToString() -> String? {
-////        let numberFormatter = NumberFormatter()
-////        numberFormatter.numberStyle = .decimal
-////        let result = numberFormatter.string(from: NSNumber(value: self))
-////        return result
-////    }
-////}
-//
-//extension WITextFieldView {
-//    public enum Unit {
-//        case won
-//        case day
-//        case none
-//
-//        var text: String? {
-//            switch self {
-//            case .won:
-//                return "원"
-//            case .day:
-//                return "일"
-//            case .none:
-//                return nil
-//            }
-//        }
-//    }
-//
-//    public enum NumberType {
-//        case price
-//        case day
-//        case nickName
-//
-//        var number: Int? {
-//            switch self{
-//            case .price:
-//                return 11
-//            case .day:
-//                return 3
-//            case .nickName:
-//                return 8
-//            }
-//        }
-//    }
-//}
+
+import Combine
+import UIKit
+
+import SnapKit
+
+public final class WITextFieldView: UIView {
+    private var type: WITextFieldType
+    
+    private let label = UILabel()
+    private let textField = UITextField()
+    
+    public var uploadPrice: ((_ data: Int) -> Void)?
+    public var uploadName: ((_ data: String) -> Void)?
+    
+    public let textFieldDidEndEditingPublisher = PassthroughSubject<Void, Never>()
+    
+    public lazy var pricePublisher = PassthroughSubject<Int, Never>()
+    public lazy var countPublisher = PassthroughSubject<Int, Never>()
+    public lazy var stringPublisher = PassthroughSubject<String, Never>()
+    
+    public lazy var bag = Set<AnyCancellable>()
+    
+    private lazy var name: String = "" {
+        didSet { stringPublisher.send(name) }
+    }
+    
+    private lazy var price: Int = 0 {
+        didSet { pricePublisher.send(price) }
+    }
+    
+    public override var intrinsicContentSize: CGSize {
+        CGSize(width: UIScreen.main.bounds.width, height: Size.height.rawValue)
+    }
+    
+    // MARK: - Init func
+    
+    public init(type: WITextFieldType) {
+        self.type = type
+        super.init(frame: .zero)
+        setTextField()
+        setLayout()
+        setLabel(0)
+        bind()
+        setAddTarget()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Methods
+    
+    private func setTextField() {
+        textField.delegate = self
+        textField.addLeftMargin(width: type.textLeftPadding)
+        textField.addRightMargin(width: type.textRightPadding)
+        textField.makeCornerRound(radius: Size.cornerRadius.rawValue)
+        textField.makeBorder(width: Size.borderWidth.rawValue, color: Color.inactiveBorder.color)
+        textField.backgroundColor = Color.backgroundColor.color
+        textField.font = type.textStyle
+        textField.textColor = Color.inactiveText.color
+        textField.keyboardType = type.keyboardType
+        textField.textAlignment = type.textAlignment
+        textField.tintColor = Color.cursorColor.color
+        textField.placeholder = type.placeholder
+    }
+    
+    private func setLabel(_ count: Int) {
+        if type.keyboardType == .default {
+            label.setText("(\(count)/\(type.textLength))", attributes: type.labelStyle)
+        } else {
+            label.setText(type.label, attributes: type.labelStyle)
+        }
+        
+        label.textColor = type.labelColor
+    }
+    
+    private func setAddTarget() {
+        textField.addTarget(self, action: #selector(textfieldDidChange), for: .editingChanged)
+    }
+    
+    private func bind() {
+        countPublisher
+            .sink { [weak self] count in
+                self?.setLabel(count)
+            }
+            .store(in: &bag)
+    }
+    
+    private func setLayout() {
+        addSubview(textField)
+        
+        textField.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+        
+        addSubview(label)
+        
+        label.snp.makeConstraints {
+            $0.centerY.equalTo(textField.snp.centerY)
+            $0.trailing.equalToSuperview().inset(type.labelRightPadding)
+            $0.width.equalTo(type.labelWidth)
+        }
+    }
+    
+    private func nameValidation(text: String) -> Bool {
+        let arr = Array(text)
+        let pattern = "^[ㄱ-ㅎㅏ-ㅣ가-힣a-zA-Z0-9]$"
+        if let regex = try? NSRegularExpression(pattern: pattern, options: .caseInsensitive) {
+            var index = 0
+            while index < arr.count { // string 내 각 문자 하나하나 마다 정규식 체크 후 충족하지 못한것은 제거.
+                let results = regex.matches(in: String(arr[index]), options: [], range: NSRange(location: 0, length: 1))
+                if results.count == 0 {
+                    return false
+                } else {
+                    index += 1
+                }
+            }
+        }
+        return true
+    }
+    
+    private func makeComma() {
+        guard let str = self.textField.text else {
+            textField.placeholder = ""
+            return
+        }
+        
+        var answer: Int = 0
+        
+        if str.count > 3 {
+            answer = Int(str.replacingOccurrences(of: ",", with: "")) ?? 0
+        } else {
+            answer = Int(str) ?? 0
+        }
+        
+        uploadPrice?(answer)
+        
+        self.price = answer
+        self.textField.text = answer.addCommaToString()
+    }
+    
+    public func become() {
+        textField.becomeFirstResponder()
+    }
+    
+    public func resign() {
+        textField.resignFirstResponder()
+    }
+    
+    public func resetPrice() {
+        textField.text = "0"
+    }
+    
+    public func makeErrorView() {
+        textField.makeBorder(width: Size.borderWidth.rawValue, color: Color.errorBorderColor.color)
+        textField.textColor = Color.errorTextColor.color
+    }
+    
+    public func makeActiveView() {
+        textField.makeBorder(width: Size.borderWidth.rawValue, color: Color.activeColor.color)
+        textField.textColor = type.activeTextColor
+    }
+    
+    public func makeInactiveView() {
+        textField.makeBorder(width: Size.borderWidth.rawValue, color: Color.inactiveBorder.color)
+        textField.textColor = Color.inactiveText.color
+    }
+    
+    public func changeTextLength(_ length: Int) {
+        type.textLength = length
+    }
+}
+
+extension WITextFieldView: UITextFieldDelegate {
+    
+    /// textFieldDidBeginEditing: textField의 편집이 시작될 때의 동작을 정의한 함수
+    public func textFieldDidBeginEditing(_ textField: UITextField) {
+        textField.textColor = type.activeTextColor
+        textField.makeBorder(width: Size.borderWidth.rawValue, color: Color.activeColor.color)
+        
+        if type.keyboardType == .numberPad {
+            pricePublisher.send(price)
+            
+            if let pure = textField.text {
+                if pure == type.placeholder { textField.text = nil }
+                else { textField.text = pure }
+            }
+        } else {
+            stringPublisher.send(name)
+            if textField.text == nil || textField.text == "" { textField.placeholder = "" }
+        }
+    }
+    
+    /// textfieldDidChange: textField의 text가 변경되었을때 작동하는 함수
+    @objc
+    private func textfieldDidChange(_ sender: UITextField) {
+        
+        if type.keyboardType == .numberPad {
+            makeComma()
+            
+            guard let text = textField.text else { return }
+            if text == type.placeholder {
+                textField.text = ""
+            } else {
+                textField.text = text
+            }
+        } else if type.keyboardType == .default {
+            countPublisher.send(textField.text?.count ?? 0)
+            if textField.text == "" { textField.placeholder = "" }
+        }
+    }
+    
+    /// textFieldDidEndEditing: textField의 편집이 종료되었을때 작동하는 함수
+    public func textFieldDidEndEditing(_ textField: UITextField) {
+        if ((textField.text?.isEmpty) != nil) {
+            textField.placeholder = type.placeholder
+            self.makeInactiveView()
+        }
+        textFieldDidEndEditingPublisher.send(Void())
+    }
+    
+    /// textField: 텍스트필드에 새로운 문자가 추가되었을때 text를 바꿔주는 함수
+    public func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange,
+                          replacementString string: String) -> Bool {
+        
+        var price: Int = 0
+        var isCorrect: Bool = true
+        
+        if let pureText = textField.text {
+            price = pureText.count + string.count - range.length
+            
+            if type.keyboardType == .default {
+                isCorrect = nameValidation(text: pureText+string)
+            }
+        }
+        return !(price > type.textLength) && isCorrect
+    }
+}
+
+
+extension UITextField {
+    func addLeftMargin(width: CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+
+    func addRightMargin(width: CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
+        self.rightView = paddingView
+        self.rightViewMode = .always
+    }
+}
+
+extension UIView {
+    func addSubviews(_ views: UIView...) {
+        views.forEach { self.addSubview($0) }
+    }
+    
+    func makeCornerRound(radius: CGFloat) {
+        layer.cornerRadius = radius
+        layer.masksToBounds = true
+    }
+    
+    func makeBorder(width: CGFloat, color: UIColor ) {
+        layer.borderWidth = width
+        layer.borderColor = color.cgColor
+    }
+}
+
+extension Int {
+    func addCommaToString() -> String? {
+        let numberFormatter = NumberFormatter()
+        numberFormatter.numberStyle = .decimal
+        let result = numberFormatter.string(from: NSNumber(value: self))
+        return result
+    }
+}
+
+extension WITextFieldView {
+    enum Size: CGFloat {
+        case cornerRadius = 5
+        case borderWidth = 1
+        case height = 56
+    }
+    
+    enum Color {
+        case inactiveBorder
+        case inactiveText
+        case activeColor
+        case errorTextColor
+        case errorBorderColor
+        case cursorColor
+        case backgroundColor
+        
+        var color: UIColor {
+            switch self {
+            case .backgroundColor:
+                return .winey_gray0
+            case .inactiveBorder:
+                return .winey_gray200
+            case .activeColor, .cursorColor:
+                return .winey_purple400
+            case .errorTextColor:
+                return .winey_gray900
+            case .errorBorderColor:
+                return .winey_red500
+            case .inactiveText:
+                return .winey_gray500
+            }
+        }
+    }
+}
